requirements.txt
beautifulsoup4>=4.12.3
requests>=2.32.3
urllib3>=2.2.2
python-dotenv>=1.0.1
openai>=1.51.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.1
numpy>=1.26.4
markitdown>=0.0.1
weaviate-client>=4.6.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
